fix: update header file: 	size_t read_a(...);     from 	int read_a(...);

+++ include/hp/hp_net.h
@@ -56,7 +56,7 @@ ssize_t
 hp_net_recvfrom_flags(int fd, void *ptr, size_t nbytes, int *flagsp,
 			   struct sockaddr_in *sa, socklen_t *salenptr, struct sockaddr_in * origdst);
 #endif /* _MSC_VER */
-int read_a(hp_sock_t fd, int * err, char * buf, size_t len, size_t bytes);
+size_t read_a(hp_sock_t fd, int * err, char * buf, size_t len, size_t bytes);
 
 #ifndef NDEBUG
 int test_hp_net_main(int argc, char ** argv);
